docs: update dataLoaderPerBatchRequest example
Hey,

I'm currently adding batch support to my own server. In order to do so I've been looking at this example. Because it seemed overly complex I decided to look into the source code and maybe simplify it a bit/clean it up.

This is what I've come up with. Instead of providing a settings creation function, storing its result to `req.graphqlServerSettings` and then restoring it from there, one can save a lot of function calls (not this big an issue to be fair) and complexity by just providing static settings and making use of `express-graphql`'s standard behavior of injecting the request as the context. Combine this with a shared `Router` for the endpoints (making sure they both get injected the `dataLoaders` and they have the same base url) and it makes for a lot easier solution.

It also has the advantage of being able to inject additional shared middleware, i.e.
```typescript
(req, res) => {
  if (req.isAuthenticated()) {
    return next()
  } else {
    res.sendStatus(401)
  }
}
```

I'm looking forward to your feedback.